+++ lib.esm/utils/geturl-browser.js
@@ -6,7 +6,7 @@ export async function getUrl(req, _signal) {
         info: { protocol },
         operation: "request"
     });
-    assert(!req.credentials || req.allowInsecureAuthentication, "insecure authorized connections unsupported", "UNSUPPORTED_OPERATION", {
+    assert(protocol === "https" || !req.credentials || req.allowInsecureAuthentication, "insecure authorized connections unsupported", "UNSUPPORTED_OPERATION", {
         operation: "request"
     });
     let signal = undefined;