+++ lib.esm/utils/geturl.js
@@ -12,7 +12,7 @@ export async function getUrl(req, signal) {
         info: { protocol },
         operation: "request"
     });
-    assert(!req.credentials || req.allowInsecureAuthentication, "insecure authorized connections unsupported", "UNSUPPORTED_OPERATION", {
+    assert(protocol === "https" || !req.credentials || req.allowInsecureAuthentication, "insecure authorized connections unsupported", "UNSUPPORTED_OPERATION", {
         operation: "request"
     });
     const method = req.method;